+++ contrib/pyln-spec/bolt4/requirements.txt
@@ -1,1 +1,1 @@
-../requirements.txt
+pyln.proto >= 0.9.3
