--- contrib/pyln-spec/requirements.txt
@@ -1,1 +1,0 @@
-pyln.proto >= 0.9.3
